requirements.txt
pytest
selene
allure-pytest
requests
